+++ tasty-plutus/src/Test/Tasty/Plutus/Context.hs
@@ -43,8 +43,10 @@ module Test.Tasty.Plutus.Context (
 
   -- ** Paying
   paysToPubKey,
+  paysToPubKeyWithDatum,
   paysTokensToPubKey,
   paysToWallet,
+  paysToWalletWithDatum,
   paysTokensToWallet,
   paysLovelaceToPubKey,
   paysLovelaceToWallet,
@@ -54,12 +56,16 @@ module Test.Tasty.Plutus.Context (
 
   -- ** Spending
   spendsFromPubKey,
+  spendsFromPubKeyWithDatum,
   spendsTokensFromPubKey,
-  spendsFromWallet,
-  spendsTokensFromWallet,
   spendsFromPubKeySigned,
+  spendsFromPubKeyWithDatumSigned,
   spendsTokensFromPubKeySigned,
+  spendsFromWallet,
+  spendsFromWalletWithDatum,
+  spendsTokensFromWallet,
   spendsFromWalletSigned,
+  spendsFromWalletWithDatumSigned,
   spendsTokensFromWalletSigned,
   spendsFromOther,
   spendsTokensFromOther,
@@ -177,7 +183,24 @@ paysToPubKey ::
   Value ->
   ContextBuilder p
 paysToPubKey pkh =
-  output . Output (PubKeyType pkh) . GeneralValue
+  output . Output (PubKeyType pkh Nothing) . GeneralValue
+
+{- | Indicate that a payment must happen to the given public key, worth the
+ given amount and the given datum attached.
+
+ @since 5.3
+-}
+paysToPubKeyWithDatum ::
+  forall (p :: Purpose) (a :: Type).
+  (ToData a) =>
+  PubKeyHash ->
+  Value ->
+  a ->
+  ContextBuilder p
+paysToPubKeyWithDatum pkh val dt =
+  let value = GeneralValue val
+      outType = PubKeyType pkh (Just $ toBuiltinData dt)
+   in output $ Output outType value
 
 {- | Indicate that the given 'Tokens' controlled by the tested minting policy
 must be paid to the given public key.
@@ -190,7 +213,7 @@ paysTokensToPubKey ::
   Tokens ->
   ContextBuilder ( 'ForMinting r)
 paysTokensToPubKey pkh (Tokens tn pos) =
-  output . Output (PubKeyType pkh) $ TokensValue tn pos
+  output . Output (PubKeyType pkh Nothing) $ TokensValue tn pos
 
 {- | Indicate that a payment must happen to the given 'Wallet', worth the
  given amount.
@@ -203,6 +226,20 @@ paysToWallet ::
   Value ->
   ContextBuilder p
 paysToWallet wallet = paysToPubKey (walletPubKeyHash wallet)
+
+{- | Indicate that a payment must happen to the given 'Wallet', worth the
+ given amount and the given datum attached.
+
+ @since 5.3
+-}
+paysToWalletWithDatum ::
+  forall (p :: Purpose) (a :: Type).
+  (ToData a) =>
+  Wallet ->
+  Value ->
+  a ->
+  ContextBuilder p
+paysToWalletWithDatum wallet = paysToPubKeyWithDatum (walletPubKeyHash wallet)
 
 {- | Indicate that the given 'Tokens' controlled by the tested minting policy
  must be paid to the given 'Wallet'.
@@ -292,7 +329,24 @@ spendsFromPubKey ::
   Value ->
   ContextBuilder p
 spendsFromPubKey pkh =
-  input . Input (PubKeyType pkh) . GeneralValue
+  input . Input (PubKeyType pkh Nothing) . GeneralValue
+
+{- | Indicate that the given amount must be spent from the given public key,
+ with the given datum attached.
+
+ @since 5.3
+-}
+spendsFromPubKeyWithDatum ::
+  forall (p :: Purpose) (a :: Type).
+  (ToData a) =>
+  PubKeyHash ->
+  Value ->
+  a ->
+  ContextBuilder p
+spendsFromPubKeyWithDatum pkh val dt =
+  let value = GeneralValue val
+      inType = PubKeyType pkh (Just $ toBuiltinData dt)
+   in input $ Input inType value
 
 {- | Indicate that the given 'Tokens' controlled by the tested minting policy
  must be spent from the given public key.
@@ -305,7 +359,7 @@ spendsTokensFromPubKey ::
   Tokens ->
   ContextBuilder ( 'ForMinting r)
 spendsTokensFromPubKey pkh (Tokens tn pos) =
-  input . Input (PubKeyType pkh) $ TokensValue tn pos
+  input . Input (PubKeyType pkh Nothing) $ TokensValue tn pos
 
 {- | As 'spendsFromPubKey', with an added signature.
 
@@ -330,6 +384,19 @@ spendsTokensFromPubKeySigned ::
 spendsTokensFromPubKeySigned pkh (Tokens tn pos) =
   signedWith pkh <> spendsTokensFromPubKey pkh (Tokens tn pos)
 
+{- | As 'spendsFromPubKeyWithDatum', with an added signature.
+
+ @since 5.3
+-}
+spendsFromPubKeyWithDatumSigned ::
+  forall (p :: Purpose) (a :: Type).
+  (ToData a) =>
+  PubKeyHash ->
+  Value ->
+  a ->
+  ContextBuilder p
+spendsFromPubKeyWithDatumSigned pkh v dt = spendsFromPubKeyWithDatum pkh v dt <> signedWith pkh
+
 {- | Indicate that the given amount must be spent from the given 'Wallet'.
 
  @since 1.0
@@ -341,7 +408,7 @@ spendsFromWallet ::
   ContextBuilder p
 spendsFromWallet wallet = spendsFromPubKey (walletPubKeyHash wallet)
 
-{- | Indicate that the given 'Tokens controlled by the tested minting policy
+{- | Indicate that the given 'Tokens' controlled by the tested minting policy
  must be spent from the given 'Wallet'
 
  @since 6.0
@@ -352,6 +419,19 @@ spendsTokensFromWallet ::
   Tokens ->
   ContextBuilder ( 'ForMinting r)
 spendsTokensFromWallet wallet = spendsTokensFromPubKey (walletPubKeyHash wallet)
+
+{- | Indicate that the given amount must be spent from the given 'Wallet'.
+
+ @since 5.3
+-}
+spendsFromWalletWithDatum ::
+  forall (p :: Purpose) (a :: Type).
+  (ToData a) =>
+  Wallet ->
+  Value ->
+  a ->
+  ContextBuilder p
+spendsFromWalletWithDatum wallet = spendsFromPubKeyWithDatum (walletPubKeyHash wallet)
 
 {- | As 'spendsFromWallet', with an added signature.
 
@@ -376,6 +456,19 @@ spendsTokensFromWalletSigned ::
   ContextBuilder ( 'ForMinting r)
 spendsTokensFromWalletSigned wallet =
   spendsTokensFromPubKeySigned (walletPubKeyHash wallet)
+
+{- | As 'spendsFromWalletWithDatum', with an added signature.
+
+ @since 5.3
+-}
+spendsFromWalletWithDatumSigned ::
+  forall (p :: Purpose) (a :: Type).
+  (ToData a) =>
+  Wallet ->
+  Value ->
+  a ->
+  ContextBuilder p
+spendsFromWalletWithDatumSigned wallet = spendsFromPubKeyWithDatumSigned (walletPubKeyHash wallet)
 
 {- | Indicate that the given amount must be spent from another script.