+++ tasty-plutus/src/Test/Tasty/Plutus/Internal/Context.hs
@@ -101,8 +101,8 @@ data Purpose where
  @since 3.0
 -}
 data ExternalType
-  = -- | @since 3.0
-    PubKeyType PubKeyHash
+  = -- | @since 5.3
+    PubKeyType PubKeyHash (Maybe BuiltinData)
   | -- | @since 3.0
     ScriptType ValidatorHash BuiltinData
   | -- | @since 3.0
@@ -374,13 +374,13 @@ toInputDatum :: Input -> Maybe (DatumHash, Datum)
 toInputDatum (Input typ _) = case typ of
   ScriptType _ dt -> Just . datumWithHash $ dt
   OwnType dt -> Just . datumWithHash $ dt
-  PubKeyType _ -> Nothing
+  PubKeyType _ dt -> datumWithHash <$> dt
 
 toOutputDatum :: Output -> Maybe (DatumHash, Datum)
 toOutputDatum (Output typ _) = case typ of
   ScriptType _ dt -> Just . datumWithHash $ dt
   OwnType dt -> Just . datumWithHash $ dt
-  PubKeyType _ -> Nothing
+  PubKeyType _ dt -> datumWithHash <$> dt
 
 datumWithHash :: BuiltinData -> (DatumHash, Datum)
 datumWithHash dt = (datumHash dt', dt')
@@ -393,7 +393,7 @@ createTxInInfo conf (ix, Input typ valType) =
   let val = extractValue conf valType
    in TxInInfo (TxOutRef (testTxId conf) ix) $
         case typ of
-          PubKeyType pkh -> TxOut (pubKeyHashAddress pkh) val Nothing
+          PubKeyType pkh dat -> TxOut (pubKeyHashAddress pkh) val $ datumHash . Datum <$> dat
           ScriptType hash dat ->
             TxOut (scriptHashAddress hash) val . justDatumHash $ dat
           OwnType dat ->
@@ -406,7 +406,7 @@ toTxOut :: TransactionConfig -> ValidatorHash -> Output -> TxOut
 toTxOut conf valHash (Output typ valType) =
   let val = extractValue conf valType
    in case typ of
-        PubKeyType pkh -> TxOut (pubKeyHashAddress pkh) val Nothing
+        PubKeyType pkh dat -> TxOut (pubKeyHashAddress pkh) val $ datumHash . Datum <$> dat
         ScriptType hash dat ->
           TxOut (scriptHashAddress hash) val . justDatumHash $ dat
         OwnType dat ->