+++ tasty-plutus/CHANGELOG.md
@@ -4,7 +4,7 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
-## 6.0 -- 2022-01-14
+## 6.0 -- 2022-01-19
 
 ### Added
 
@@ -49,9 +49,17 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
   * `mintOutcome` to `mpOutcome`
   * `mintTokens :: Tokens` to `mpTasks :: NonEmpty MintimngPolicyTask`
 
+## 5.3 -- 2022-01-17
+
+### Added
+
+* Added `paysToPubKeyWithDatum`, `paysToWalletWithDatum` including a Datum for pub key outputs and
+  `spendsFromPubKeyWithDatum`, `spendsFromPubKeyWithDatumSigned`, `spendsFromWalletWithDatum`,
+  `spendsFromWalletWithDatumSigned` for pub key inputs
+
 ## 5.2 -- 2022-01-10
 
-### Changes
+### Changed
 
 * `toTestValidator` and `toTestMintingPolicy` now accepts any `FromData` as the script context.